Showing errors and success messages
Showing errors and success messages in admin part of app

+++ app/Http/Controllers/User/DeleteController.php
@@ -9,7 +9,8 @@ class DeleteController extends Controller
 {
     public function __invoke(User $user)
     {
+        $this->authorize('delete', [$user, auth()->user()]);
         $user->delete();
-        return redirect()->route('user.index');
+        return redirect()->route('user.index')->with(['success' => 'Пользователь id' . $user->id . ' успешно удален']);
     }
 }
